Permit whitespace between colon and name in label
A side effect of not allowing whitespace here is that whitespace is
required in casting expressions in order to disambiguate. For example,
a:int would lex as name a, label int. The simplest fix for this is to
handle labels during the parsing stage instead of the lexing stage,
however, this allows whitespace between the colon and the name. This
shouldn't introduce any ambiguities in the grammar, so it seems
reasonable to allow it.

Signed-off-by: Sebastian <sebastian@sebsite.pw>