Require size to include zero.
A slice is equivalent to a struct including a field for its length, of
type size. Because the length of a slice can be zero, the size type must
be able to contain the value zero. The size type is also unsigned, so
the minimum value cannot be less than zero. Thus the minimum value of
the size type must be zero.

Before this patch, an implementation could define the minimum value of
size to be a positive integer (e.g. one), preventing the existence of
empty slices.

This argument also applies to the string type.

Signed-off-by: Chloe Brown <chloe.brown.00@outlook.com>