Fix typos in 6.6.15.8, 6.5.20.4, 6.6.47.1, 6.6.47.2 and 6.6.49.3.
Proposed fixes for some inconsistencies I have found in the specs.

In 6.6.15.8 the size of a struct should the sum (not the max) of its members.
In 6.5.20.4 a "parameters" should be "parameters-list".
In 6.6.47.1 "expression" is repeated twice.
In 6.6.47.2 "binding-list" should be "binding".
In 6.6.49.3 "an" should be "a".

It seems to me that most of them may have been generated by changing the
schematic part of a section (i.e. splitting something in more than one
line for clarity), but not the parts of the text that referred to it.
The only exception is the struct size one, which seems a copy-paste
mistake.

Signed-off-by: Sebastiano Tronto <sebastiano@tronto.net>