Specify test environment
This also more concretely specifies behavior of initialization and
finalization environments.

Specifying this as its own environment separate from the hosted
environment is necessary, since the test environment doesn't treat main
as the program entry point. Furthermore, the behavior of abort is
different in the test environment. So without specifying this, the test
runner is non-conforming. The test runner is fairly straightforward to
specify, and doesn't impose any additional burden on the implementation
since support for it is optional.

Signed-off-by: Sebastian <sebastian@sebsite.pw>