Loosen definition of "undefined"
"Programs which rely on these details are non-conforming."
This is unnecessary, since a lot of undefined details can't be "relied
on", since there's nothing that *can* be relied on. Like, you can't rely
on an undefined default value or undefined size (you could try to take
the size of a type whose size is undefined, but that's just an invariant
being broken). Also, if there's UB at runtime, the rest of the program
should still run as normal, with caveats for what's unpredictable (in
some cases, only a value is undefined, rather than all behavior).

Signed-off-by: Sebastian <sebastian@sebsite.pw>