Disallow consecutive commas in constant-bindings and global-bindings
Previously we had the rules:

1. <global-bindings> ::= <global-binding>
2. <global-bindings> ::= <global-binding> ","
3. <global-bindings> ::= <global-bindings> "," <global-binding>

However, this would allow:

<global-binding> "," "," <global-binding>
├─<global-bindings>┘                    │  By rule 2.
└─<global-bindings>─────────────────────┘  By rule 3.

which is undesirable (and doesn't match harec behaviour).

./demo.ha:1:11: syntax error: unexpected ',', expected ';'

1 |     let a = 1,, b = 2;
  |               ^

We can fix this by swapping the order of rule 3, so that the recursion is on
the right. This also matches the behaviour of struct-union-fields, case-options,
bindings, etc, all of which recurse on the right.