allow comments to be terminated with eof
simpler than the alternative of requiring a trailing newline everywhere,
and matches harec's current behavior. note that comment-char is defined
as any character other than \n, so just removing the \n from comment is
enough

one argument against this is that it doesn't match posix's definition of
text files, but this doesn't really matter all that much, and not
requiring a trailing newline is consistent with eg. clang's and gcc's
behavior

Signed-off-by: Ember Sawady <ecs@d2evs.net>