Replace scalar types with primitive types
There was section which said "The type class of a type is defined for
scalar types as the terminal which represents it". This is incorrect,
since pointer types were also considered scalar types. I've moved
pointer types out of this part of the grammar, and renamed it to
primitive-type. The spec already uses "primitive type" in the section
"Other primitive types", so this is consistent with that. This also
allows str to be added as a primitive type, meaning the string-type
"nonterminal" (which is actually just the terminal str) can also be
removed.

The pointer types section is also moved to after "Other primitive
types", since pointer types aren't primitive types.

Signed-off-by: Sebastian <sebastian@sebsite.pw>