Improve description of struct/union default value
This description mentioned overlapping fields in structs, which are now
no longer possible. A later section on struct/union equality notes that
equality doesn't depend on the order of the fields in the program
source, and I've kept this note since 1. unions exist, and 2. two fields
may still share the same offset, if at least one of them has size 0.

I've also added that the default value of unions is always undefined.

Signed-off-by: Sebastian <sebastian@sebsite.pw>